+++ tests/bsim/bluetooth/ll/bis/prj_past.conf
@@ -1,0 +1,33 @@
+CONFIG_BT=y
+CONFIG_BT_DEVICE_NAME="PAST"
+CONFIG_BT_PERIPHERAL=y
+CONFIG_BT_CENTRAL=y
+CONFIG_BT_EXT_ADV=y
+CONFIG_BT_PER_ADV=y
+CONFIG_BT_PER_ADV_SYNC=y
+CONFIG_BT_CTLR_ADV_EXT=y
+
+CONFIG_BT_CTLR_TEST=y
+
+CONFIG_BT_ATT_PREPARE_COUNT=2
+CONFIG_BT_GATT_CLIENT=y
+CONFIG_BT_L2CAP_DYNAMIC_CHANNEL=y
+CONFIG_BT_DEVICE_NAME="bsim_test_split"
+CONFIG_BT_L2CAP_TX_BUF_COUNT=6
+
+CONFIG_BT_CTLR_PRIVACY=y
+
+# Periodic Advertising Sync Transfer Send/Receive functionality
+CONFIG_BT_PER_ADV_SYNC_TRANSFER_RECEIVER=y
+CONFIG_BT_PER_ADV_SYNC_TRANSFER_SENDER=y
+
+# Increase buffer size to get whole LLCP message sent in one
+CONFIG_BT_BUF_CMD_TX_SIZE=255
+CONFIG_BT_BUF_EVT_RX_SIZE=255
+CONFIG_BT_BUF_EVT_DISCARDABLE_SIZE=255
+CONFIG_BT_BUF_ACL_TX_SIZE=251
+CONFIG_BT_BUF_ACL_RX_SIZE=251
+
+CONFIG_BT_CTLR_DATA_LENGTH_MAX=251
+
+CONFIG_BT_BROADCASTER=y
